move controls to new file
